feature: Make Overdraft disjoint and independent of Money
This makes some breaking changes to how Overdraft works.

Overdraft no longer accepts a value of zero, but instead raises on
instantiation. This includes a revert of 7511220, which made Pydantic
behavior inline with previous behavior of accepting zero as value for
both Money and Overdraft.

Overdraft no longer has a `.money` attribute, but instead now carries
its own `.value` and `.currency` attributes.

`MoneyParseError` is renamed to just `ParseError`, because it is also
raised on instantiation of `Overdraft`. This was the case prior as well,
but it was raised because an instantiation of an `Overdraft` object
always also instantiated a `Money` object.