feature: Support pydantic-core 2.10+
pydantic-core 2.10 deprecated some functions in the
`pydantic_core.core_schema.*` namespace. In order to both be compatible
with future (no deprecation errors), and have working type checking,
there's no choice but to drop support for pydantic-core versions <2.10.

Because this drops support for some pydantic-core versions, it is a
breaking change.